Finish main code restructure
